dapp: add regular cache integrity check
The service worker is already able to recognize that the cache is
broken. But it does so only when there is an error while serving from
the cache. While important to have it, it has some weaknesses. There
could be the case that the cache is partially broken. This means the
user can continue the application until he enters/access something that
is broken. This could happen in an disadvantageous time. E.g. when he
entered a critical phase. We can't really know what issues such a "gap"
in the cache could cause. Therefore it is saver to detect such cases
early and prevent the user to do anything anymore.
To do so the assistant regularly asks the service worker to verify the
cache validity. If it is not valid anymore, the service worker will
trigger a mandatory update.
After all there is still a (small) time gap in which the cache can get
broken and the user continues the application. The gaps size is defined
by how frequent the check is done. Thereby it is a balance of doing not
too many checks to prevent a high load on the resources and doing it
often enough that it is unlikely that the user can get into trouble
(which is still a vague expression).