dapp: add install banner component
The application is finally able to be installed as progressive web app.
On desktop there is a tiny plus sign in the URL bar and on mobile
a pop-up gets shown. This allows the user to "install" the application
to the desktop or home screen, makes it feel more like a native client.
To advertise this feature more and make it actually visible on desktop,
this commit adds a new component to show an installation banner at the
top of the page. This makes use of the special window event that gets
triggered if the application is recognized as installable.
If the user dismiss the banner, the banner will be shown after the next
reload again. Here it is the same as for the update snackbar: it is just
more work to store this decision correctly (and decide when to show
again) combined with we want the user to not ignore it.

The "special" event is not natively supported by the TypeScript
libraries. Also it does not work in all web-browsers (here it just do
nothing). Therefore I had to add these types manually.

During the installation tests I discovered that we use a horrible color
for the theming in our PWA manifest. This color is not even part of our
palette. So I changed it to our primary color.