dapp: reduce unnecessary install of service worker
The code to register the service worker get executed each time the
applications loads. No matter if there is already a service worker or
not. Most of the time this does nothing since the browser detects that
the service worker script are identical. But in case there is an update
at the service this must not lead to install the new sevice worker. The
old should just continue to run.
Therefore the plugin now checks if there is already a controlling worker
available. For an initial load or an update this will still register the
service worker, else not.