dapp: add service worker database functionality
Each newly installed service worker must detect if there is already
a locally cached version or not. If so it must continue serving this
version. The so far overseen issue is that the service worker re-uses
the old cache, but the asset list it is maintaining does not match. Each
servic worker gets bundled with a static list of to precache assets.
This list is not only used to perform the initial precaching, but also
for the routing. Means the router checks if the request matches an
assets it is controlling. This for exmaple allows the user to still
access the docs that lies outside of the application the service worker
controls. If the new service worker just use its new static list to
match requests, it will eventually not link correctly with the cache.
Even worse, the service worker will detect that the cache is broken from
its point of view and force the user to update.
To solve this issue, each service worker that continues serving a local
version must also use the precache asset list of that version. Therefore
the service worker who is doing the precaching must not only download
the assets to the cache sotrage, but also preserve this list. Since the
cache storage can only save reponse objects, this requires the usage of
the browesers indexdb API (remind that there is no access to the local
storage API).
This comes with a new special case that a new service worker founds an
old cache, but not a corresponding asset list. This gets detected and
will force an update. The opposite case doesn't matter, as without any
cache the worker will go for an update that finally overwrites the asset
list in the database.

During this commit there was the need to start splittin the utilities
module of the service worker directory. It started to become too much
cluttered after adding the database functionality. Therefore got all
functions separated by the browser API they are using (i.e. cache
storage, indexdb, clients). The index file is then describing the hight
level logic by orchestrating all of the utilitiy functions.