dApp: fix flickering of ongoing transfer snackbar
Transfers in Raiden are usually blazing fast. This lead to the issue
that the snackbar to inform the user about an ongoing transfer was
"flickering". Should mean it was displayed for such a short period
(sub-second) that it was more confusing rather helpful.
As work-around we agreed to lengthen the time the snackbar gets shown,
even the transfer is already finished. Such a solution must make sure
that there could be multiple ongoing transfers that get received. Also
the first transfer could be done and then the next starts while the
snackbar is still "debouncing". Here we must make sure that the snackbar
continuous showing until the end of the last transfer.
After some complicated approaches to handle this with RxJS, the most
simple solution is to use a CSS transition fade-out effect. In fact this
simply delays the default effect (just remove the component...). This
solution is secure against the case that if while this delayed fade-out
transition is ongoing, it is fading in again, this will cancel the
fade-out. Simply said the component remains visible.