dapp: fix back-button navigation to remember token
The back button of the header component leads to the issue that if the
user changed the token network, navigates around (e.g. open channel
list) and then comes back to the transfer route, the selection is
forgotten.
This is caused by pushing directly to the router without preserving any
route parameters. Therefore this could been solved by easily tweaking
the router usage by the button.