dapp: fix persistence of disclaimer acceptance
If the user accetps the disclaimer and also selects to remember this
decision, the disclaimer must be skipped from then on. This feature was
broken due to the backup notification feature. That was mixing the
disclaimer persistence plugin with the new backup reminder state. Means
the values get persisted to local storage when any of both parts have
changed. The value for the disclaimer gets saved as it is only if the
user decided to remember the decision. Else we write the default false
value. This works correctly the first time the user accepts the
disclaimer. When the the backup notification reminder also causes
a persistence call, this still works, since the state that the user
wants to persist this is still present. But the next time the app gets
loaded and the backup notification triggers the write to local storage,
that is not known anymore. Means it will overwrite the true with
a default false value. That will then trigger to show the disclaimer
again on the next run of the application.
The simple solution is to separate both state parts into different
persistence plugins. They should simply not get updated together. And
it does also not make sense to persist the persist decision (reads
already stupid).