sdk: add origin parameter for subkey generation
The user must be able to recover state backups of service providers that
are not available anymore. This requires to make the subkey
configuration configurable to overwrite the origin URL explicitly. This
then allows the user to run the service locally and specify the URL of
the offline service to get the same subkey that gives him access to the
backup.

The feature got established by a new parameter to the SDK
instantiation. This keeps the origin URL specification close to the
parameter that defines to use a subkey or not. Alternatively this could
have been put into the configuration object. The disadvantage would be
first the negotiation of the former advantage. And as the subkey
parameter, this is not a configuration option at a higher level, but
only used inside the helper function that creates the SDK instance. This
get obvious when you try to add this as a configuration option as the
code becomes chaotic.