dapp: add env variable for backup recovery
To recover a backup of a service provider that is not available anymore
it requires to overwrite the origin URL. Therefore the user can specify
this variable for serving the dApp locally. The now generated subkey
will then be equal to the one that got generated for the offline
service. Finally the user is able to import his old backup again.