dapp: fix insecure window reload of sw assistant
When the service worker assistant receives the message to reload the
window, it was waiting 1s until doing so. This was based on the wrong
assumption, that the service worker must get some time to unregister
first. This again was based on the assumption that the service worker
can't send messages to the client anymore after unregistering itself.
Since this is wrong, the worker can first unregister itself and then
send the message to reload the window. Thereby the assistant must not
set a random timeout where he never knows if it is correct or not.

Note that the unregistered service worker switches into the redundant
state. Which means he is still controlling the client, but will be
replaced by the next installed service worker.