Fix CircleCI caching issues for yarn
For some reason we got permission issues when unarchiving the cache for
the fallback cache for Yarn. Therefore we are bumping the cache version
to force the creation of a new cache. Since we already tried to fix an
issue once using version 3, we have to use 4 as version 3 is also
corrupted.