dapp: disable sw assistant when no sw is available
When the browser does not support service workers, no worker gets
registered. But the service worker still expects that it is supported.
Thereby it throws errors and eventually will show updates to the user,
even he can't update this way.
Therefore the assistant must be still initiated to provide the required
interface to the components, but it basically does nothing. No checks
for update, no message handling nor does it anything when the update
function gets called.