dapp: fix sw to take over local version of dApp
We have encountered an issue of the dApp when a new service worker is
taking over an old local version. For example if the old version has
a route that does not exist on the build where the new worker comes
from, this route will hang the first time loading. There are no errors,
no network requests or anything else. The route simply gets not
navigated to and the page is blank inside the main route view component.
After the next reload of the application everything works as expected
and the route that still exists in the old local version gets served.
Unfortunately we were not able to determine the root cause of this issue
to fully understand it. But since the page reload is a reliable fix, we
implemented that into code. This means if a new service worker gets
registered and it finds an old version, during the activation it will
reload the application another time. Since such a service worker is
super fast installed and activated as there is nothing to cache, this
reload happens incredibly fast. The user probably not even notices it.