sdk: add rx complete utils and wait epics complete
Up to now, we forcefully unubscribed epics as soon as a shutdown action
went through. This didn't give them chance to complete their tasks
gracefully. Now, the shutdown action completing the input (action$,
state$) indicates to the epics that they should complete. If they don't
in a httpTimeout, the root epic will force-unsubscribe and log a warn.
Database teardown is moved to the root epic, in order to be executed
only after all epics completed. rx utils are added to make it easier for
epics to complete when the input completes.