sdk: ensure caps are merged on defaultConfig
If a 'caps' partial config was passed to the partial config param of
Raiden.create, it'd replace entirely the default caps, forgetting values
which were omitted. This commit ensures both default default caps, and
caps overwrites passed to Raiden constructor are merged, unless
caps=null, in which case, it takes priority and disables caps entirely.