sdk: ensure caller connection is cleaned
So, in case WebRTC caller managed to send an 'offer' to a peer which
appeared offline, this messageSend.request wouldn't succeed and the
[connection, channel] pair won't hit listenDataChannel$, responsible for
closing it on 'finalize'. This commit ensures this connection is
properly cleared up on shutdown, allowing the CLI to exit cleanly even
on this case. It also fixes some unneeded decode log calls for wrong
'type'.