dapp: improve testability of sw assistant
The service worker assistant makes use of intervals that trigger
asynchronous functions. This combination is hard (not possible?) to test
with Jest its fake timer environment. Therefore it is necessary to
actually run the intervals in real time to test them. As the current
interval timeout values are partially very high (i.e. 1h), this does not
work for making tests. To fix this the interval timeouts became
configurable during instantiation. The constructor parameters default to
the original time values. This allows the tests to set very short
time values to not delay the tests, but also not hit the maximum
timeout.

Note that all used handler/callback functions of the intervals are
defines as being asychronous, also they might not need to. This has the
simple advantage that we can await all "open" promises in the tests.
Thereby we can make sure that all possibly still running handlers that
have been triggered by an interval have completed.