dapp: add unit tests for the sw assistant
Note that majority of preparation for these tests go into the handling of
intervals as we are required to actually run the intervals du to the
asynchronous handler functions.
Furthermore we need to mock the communication with the controlling
service worker. Therefore we make use of the EventTarget interface. As
there is no implementation for it in the current Jest environment we are
using, a simple implementation has been added to the Jest setup. The
implementation is a simplified version of the examples in the MDN docs
for this interface.