dapp: add check for required browser features
The dApp, but also the SDK require a couple of features of the
environment (e.g. IndexedDB). It has been decided that we don't want to
handle all eventual scenarios of missing features and partial support.
Also do we not want to check which browser engine and version supports
which feature. Rather we check globally if all features are available
once. If not, the whole application gets blocked. This works similar as
if no JavaScript is available.
There this logic goes directory into the index page and unregisters the
application in advantage of an error message if something is missing.
The list of features might get extended over time.