sdk: remove legacy state migration [BREAKING]
The new data persistence architecture with a database is used for
multiple releases already. Therefore it is time to remove the logic to
import and migrate states from the localStorage of old versions. This
cleans up the code of the SDK again and makes it better to maintain.
A user who is stuck on a very old version needs to checkout any of the
versions in-between to do the migration, before he goes for the latest
release.