cli: adopt to new SDK creation interface
The SDK does not support the legacy localStorage anymore. This has
changed the interface used in the CLI.
Since we do not create the localStorage anymore, the related data
directory does not get created anymore as well. But the database should
get stored inside and will fail if the directory does not exist. So it
still needs to get created.

As the filesystem package its promises function are not well suited to
check directory existence and create some, it got adopted to the
"regular" functions set. Instead of using promises where not needed, the
synchronous equivalents do a better job and remove some asynchronism.