dapp: fix action buttons for channels on mobile
While the buttons looks perfectly fine on desktop they are completely
out of shape on mobile. This comes in first place from some "magic" by
the used Vuetify component. As we don't need any of its fancy features
and it just brings overhead, noise and this issues, it is easier to just
use a plain HTML image instead.