dapp: fix layout issues due to mobile keyboard
When the user opens a virtual keyboard on a mobile device, this changes
the size of the viewport. This leads to a broken layout, as the whole
content does not fit into this box anymore. Apparently this does not
apply to the hidden route view (application content). Thereby its
content shines through the account while when scrolling. This got fixed
by using a proper container for the account route.

Furthermore the grid at the RaidenAccount route to transfer ETH between
the accounts got a little improvement on the grid. This makes it look
not so compressed on mobile, while preserving the desktop version.
Additionally did the bottom action button to trigger the transfer have
become sticky. This looks a bit more neat and makes the layout more
consistent. It only works/makes sense as long as the above content is
size limited. As it fits for the current content, this avoid ugly
desktop versus mobile margins/paddings to remove the scrolling from
mobile, where the button gets hidden.