dapp: reduce account menu size to fit screen
At the Chromium browser engine the account menu does not fully fit the
height of the window. This was not visible before, as the bottom margin
just disappeared and all content was visible. But do the automatically
applied scrolling, this now looks ugly as we get a scrollbar although
the content actually fits. To resolve this on desktop where the size of
the box is fixed, the margins between the menu parts got slightly
reduced.