dapp: adjust with of account menu on mobile
The margin for the account menu does work well for view desktop but not
for mobile. Due to the smaller screen width the menu looks to narrowed.
Therefore the margin got reduced on mobile devices.

Furthermore the account details table for the address and balances has
breaking lines in their header column cells on mobile. Therefore the
first column must be slightly wider on mobile devices to have a clean
table.