dapp: fix router errors due to bad navigation
Vising the transfer route without any path or query parameters led to
a router exception due to duplicated navigation. The root for this issue
was that the TransferInput component was declaring an empty token
address as invalid. In such case it was automatically redirecting the
user to the home route. As the home route gets automatically skipped if
the user is already connected, this lead to a navigation from home to
transfer to home to transfer again. This is clearly a bad behavior.

To resolve this issue, the handling of the path and query parameters
must be done in a positive manner without any automatic redirection if
anything is not fine. The first important change is that an empty
token address as the path parameter is not an error, as this is the
default and leads to a selection of the first available token by the
user. Second, the parameters should not be pre-evaluated. The parameters
get inserted into the address and amount input components. Those include
all the smart logic to evaluate the values. If any parameter is not
correct, these components will show the warning messages as for a normal
user input. Thereby no parameter gets just ignored and eventually
undetected by the user who just clicked on a link to do a transfer. This
could lead to unexpected behavior like a transfer in the wrong token
network. After all this just means that the path and query parameters
should be inserted straight, reducing the complexity and code
duplication.

Furthermore, only route components should handle the path and query
parameters themselves and pass them downwards the component hierarchy.
Therefore we use synced property values. Changes by the user are synced
upwards and changes to the route are synced downwards.

Another undetected issue was the loading of not yet known tokens. Why
this might be not a real world scenario, it is definitely a code issue.
If the token is not yet known we just pass the address to the
sub-components and then load the token model via the SDK. But if that is
done, we never update the token property. This led to some overload of
handling the token property in the children to access the decimals etc.
To make this secure, changing the token always requires the full token
model to get loaded first, if it does not exist yet. This makes all
sub-components way more simple and straight forward.

Finally a couple of minor code improvements for the readability got
applied. Such are more detailed names, better class component sorting
and single abstraction level functions.