dapp: add script for capacitor initialization
As till now we want to keep the generated Capacitor code out of our
version control system, it is required for each developer to initially
generate this himself before he tries to serve or build the mobile
application. As this is a part of the vue-cli plugin generator script,
this needs to be added in a custom implementation. Therefore a new
package script got added to initialize a certain platform if it isn't
already. This script gets automatically called by the serve and build
scripts as a convenience feature.

Note: This does not work 100% yet. While it should work fine for iOS (no
MacOS user here) there is an additional step required for Android. There
the file content of a specific generated file needs to be adopted. It is
not yet clear how to do this in a clean manner.