dapp: fix Capacitor initialization for Android
Unfortunately it requires to adopt the generated Capacitor native code
to make the development server working for Android. This originally gets
done by the Vue app, but as we don't want to re-setup everything for
each developer, we need to do this manually. Therefore the
initialization scripts for both platform can't remain fully generic, but
must be specialized for Android only.