sdk: replace messageGlobalSend with messageServiceSend actions
So far, services messages were sent to the respective global rooms (and
the epics which emitted those actions knew about it). This commit
prepares to remove those rooms, so the actions just specify to WHICH
service the message should be sent to, and the transport is responsible
for knowing about that and how to get the messages to them.