devops: rework e2e environment script to help CI
In our CI pipeline it is the goal to have the end-to-end test run
outside of the end-to-end Docker to make the workflow more clean without
redundancy of installation and build jobs. Rather the container runs in
parallel to the main executor image, having access to all the provided
services needed to the tests.
The thereby resulting issue is that there is no direct access to the
deployment information files, nor the container itself (as we do it for
local end-to-end test runs). Thereby it is required to store the
deployment information files outside the container permanently and put
them under version control. This way the main executor in the CI
pipeline will have direct access to them.
This strategy requires and additional step for each new build of the
Docker image to extract the files. Simply documenting workflows is
always error prone. Therefore a new script has been added to do the
whole build. As result of the strategy, also the script to run the test
had to change.
Both scripts share a lot in common of variables, functions and shell
settings. So despite these are "just" bash scripts, the common parts got
exported into a shared script that acts like a to import lirbray. This
way the actual to run script remain clean and focused. Furthermore
nothing is redundant which could lead to unexpected scenarios/errors.