devops: adopt CI e2e executor to stack images
So far there were two distinct executor in the CircleCI pipeline.One as
the base one, which is used for the majority of all jobs. And another
for the end-to-end tests. This has the issue that the installed
dependencies and build packages from the jobs with the base executor
can't be re-used in the end-to-end executor. Attaching the workspace
from one image to the other leads to failing tests. The root cause could
never be identified. This leads to the issue that all shared jobs must
run twice, once for each exectutor. It would be the preferred solution
to re-use the results of the shared jobs without runnimg them twice.
To do so, the base image where the jobs run inside must remain the same.
This gets possible with a feature that CircleCI calls image stacking.
Herefore a single executor runs multiple Docker container in parallel.
The jobs run only in the first container. All other containers are just
meant to provide services, available to the first container. This is
100% the use-case for the end-to-end tests.
The reason why this approach wasn't used before were issues with the
access of the deployment information files in the end-to-end environment
image. Those got resolved recently, so it is now possible to run the
end-to-end tests as they would be run on a developers localhost.