devops: adopt e2e run script to detect CI
Currently the package scripts to run the end-to-end tests with Docker
can't be used within the CircleCI jobs as it is not possible to run a
Docker container inside a container itself. But the CircleCI executor
already runs the necessary container with the necessary end-to-end
environment. So in fact this step can be just skipped.
Doing so makes the script re-usable in the CI pipeline as well. This
keeps the end-to-end test related jobs more simple and resolves any
issue due to redundant definitions and other problems. This was for
example the case for the recent changes of the deployment information
files which need to be sourced.