devops: move CI nightlies into other workflows
So far the end-to-end tests of all workspaces have been only run by a
special workflow which runs once each day at midnight. It was done this
way as the end-to-end tests take quite some time and should not lower
the productivity on daily PR work. But more than once we experienced the
issue that we introduced bugs which when got detected at night, after
the PR was already merged. Sometimes it wasn't that easy to trace back
issues to buggy PR changes. To prevent such issues the end-to-end tests
should run in the default workflow. Thereby they will also run for each
PR commit.
The total time of the default workflow actually depends on the current
SDK integration tests. Thereby the time argument gets negated. Also
these tests are target of future improvements as well as other time
saving logic.