sdk: refactor logToContractEvent for better type-safety
TypeChain now includes better events typing in filters. It's a little
bit cumbersome to extract them, but they're there. This commit leverages
that to narrow as strictly as possible the return of the
`logToContractEvent` helper function in order to use that typing on the
outputted tuple. As it comes from 'ethers', some casting is still
needed, e.g. 'string' to 'Address', but it's safer than the fully blind
type declarations we had before, and also includes more information
about those events, as struct property names.