devops: change CI e2e image for new version
Finally we are using versions for the new end-to-end environment images.
This allows us to do updates on the image and test them on this branch
without affecting other PRs and build. This plus many more advantages.
The image got just changed the commit before and should be used now.
Future changes should bump the version for both.