dapp: fix navigation loop when adding new token
The following scenario was broken: the user wants to connect a new token
and continues until the screen to select a hub. If he now wants to beak
up and clicks on the back button, he will get to the select token
screen. But if he click again on back, he will land on the select hub
page again. This loop is endless.
The cause of this loop are explicit navigations mixed with plain
"navigate-back" commands for the back button. The navigate-back was
introduced to easily preserve the selected token on the transfer/home
screen. As we need to preserve this, it is required to get rid of the
direct navigations for this case. This should be fine as we now get the
navigation path of transfer route, select token, select hub, open
channel. From any of these routes you can go just back in traverse
manner without loosing any route parameter information. If the user
finishes to open a channel, it will automatically take the shortcut to
the transfer route. As this one has no back button anymore, then
navigation path is finished. Should mean he can't go back to the open
channel route again unless he initiates this on purpose.
Direct navigations via the browser and also its navigation stack can
cause issues. But this is out of our control and expected. Though it is
never a security risk and all navigation guards do still apply.