devops: add actual check for e2e deployment info
So far, when running the end-to-end tests, it was only warning that the
deployment information files must match the used environment image. To
make this more stable, the build script is creating another file to save
the related version. When running the tests, this can be used to compare
with version inteded to run with.