dapp: extend transfer route for change detection
The following scenario has been tackled: The dApp gets started with a
new and fresh account who has no channels on the blockchain. As result,
after the user got connected, the NoTokens component gets shown on the
transfer route. If now another Raiden node opens a channel to this new
account, the transfer route switches into the mode "no-open-channels".
The reason behind this is, that the component correctly reacts to the
event that the user now has a token with an open channel. This
automatically disables the NoTokens component and shows the regular
transfer components with the history. But because the components token
selection is still empty (because there was none to select on start),
there are no open channels for this "token".
To resolve this, the state change from no token with channel to a any
token with a channel must be detected an trigger the selection of the
first (and only) token. This is kind of a delayed auto-selection of the
token. It would be the same result if the user would have reloaded the
page manually.