cli: add route for user deposit interactions
The Python client has added support for interactions with the user deposit
contracts. The SDK of the light client was supporting such for a long while
already. Though the CLI was not exposing this functionality. To align with the
interface of the Python client again, these endpoints need to be added.
This first commit just adds a new route for the purpose of user deposit
interactions. The only endpoint it implements is to retrieve information about
the current withdraw plan of the user. This endpoint is actually not part of the
Python API specification. But as the CLI has still no proper test setup, this
information is very helpful for the upcoming to implement endpoints. Therefore
it is just an additional endpoint that does not break any compatibility. But we
need to keep track of the Python client its API to see if they add a similar
endpoint and then adopt to their specification.