cli: add user deposit endpoint for withdraw plans
This commits adds the single endpoint the Python client exposes for user
deposit interactions. Though it is only a single endpoint, it provides
three different interactions possibilities. Which interactions should be
executed depends on the bodies payload.
This approach by the Python client is not RESTful at all. Having to
artificially split a single endpoint into multiple is kind of a
nightmare. But we have to align with their API, so it is required.
This commits handles in first case the challenge to provide this parsing
and decision making. It was a high goal to keep the code very simply by
having each function to deal with a single layer of abstraction only. As
downside, these three endpoints into one bring some large overhead just
to keep the route clean and readable.
Though all interactions (deposit, plan withdraw and withdraw) are fully
working, their implementation is only rudimentary. This means that if
anything is wrong with the parameters or similar, we will just response
with an internal server error including the detailed message. This will
be improved in the next commits, catching user mistakes early to provide
diverse and meaningful error codes and messages (according to the
API specification). To keep the commits reasonably small and focused
this is a wrap here.

There is at the moment the issue that we have no direct way to respond
the user with the block number when a withdraw plan is ready. There is a
hacky solution implemented that kinda works. But we need to clarify how
the SDK can provide this information to the CLI.