cli: disable auto user deposit withdraw
The light client's SDK provides both features: automatically withdraw
from the user deposit as soon as a withdraw plan becomes ready, or let
the user manually do the withdraw. The Python client provides only the
manual procedure. To align with this behavior on the CLI side, the SDK
will be configured into the manual withdraw mode.