sdk: fix an edge case on takeIf util and add tests
So `takeIf` had a weird edge condition when source$ would emit
synchronous items at subscription time, which could make it lose items
because of a race between it and a synchronous cond$ emit. Using
completeWith the output after repeatWhen caused multiple subscriptions
to the source$ as well. The fix is to replace the takeUntil+repeatWhen
combo with a switchMap completed from itself or a subject mirroring
independent source$ completions. Synchronous and assynchronous unit
tests were added to ensure this ordering is kept.