sdk: ensure promise is created soon to wait for settleChannel
The time between previous closeChannel and settleChannel was quite close
to settleTimeout, making the test flaky due to possibly partner settling
before we could wait for it. By creating the settleChannel promise as
soon as closeChannel resolves, even though it's going to be awaited only
later, ensures it'll catch any settlement which raced with us.
Also, jest seems to overwrite/ignore `process.on('beforeExit')` hooks,
so the previous workaround for wrtc segfault doesn't work. The fix is to
move wrtc polyfill on tests to rtc-specific transport tests only,
instead of globally, so webrtc is disabled on e2e tests and segfault
prevented.