dapp: upgrade Capacitor to version 3
This upgrade is necessary for the following issue. When we uploaded the
first internal test releases to the Google Play Store, we could not. The
reason was that the app requires sensible permissions to get camera and
microphone access. But the dApp is actually only using the camera.
Though the origin issue comes from the Capacitor core package. It does
not allow for treeshaking and comes with many components we don't use.
But these component ask the user for certain permissions when used. When
Google scan the application bundle it detects these refernces. And for
sensible permissions the application must come with a data security/data
protection document bundled with it. We don't have such yet. But we also
only need the camera as a nice to have feature at a later point in time.

For first experiments it was enough to change the generated Capacitor
code to remove the unwanted permissions. That made a first internal test
release of the app possible. But it is still the goal to prevent us
having any manual adoptions of the generated code. We don't want to get
into this hell. Therefore I was searching for alternatives. And the
alternative is Capacitor version 3. It has purified the core module and
outsourced everything else to official plugins that can be added when
necessary. Thereby the only permission that gets requested is the
necessary internet connection.

It must be noted that version 3 is still in an early stage on not
stable. As their is no obvious other solution to the problem and first
tests have shown that it works, this is still the solution we go for. It
is eventually not a bad idea to take the most recent version in this
early stage of the mobile app bundling. So we don't set up everything
for the "old" version and then have to update later with possibly much
more work.

The Capacitor JSON configuration file appears to get touched now and
then by Capacitor. While doing so it re-writes it in unformatted manner.
Therefore this commit includes now the single-line version of the
object.