dapp: add fastlane versioning for Android
Per default, Gradle will just using version `1` to build the Android
app. This is an issue as is, but also the Google Play store denies to
upload a second bundle with the same version. As we still don't want to
manually edit the generated Android code by Capacitor, the versioning
must be handled within the Fastlane configuration.
As solution it uses a new environment file which defines all different
version names and codes. The advantage here is that it can be easily
changed/controlled. But the major reason behind it is that we can use
plain shell commands to read/evaluate data to define them. Like reading
the version codec from the `package.json` file.
Finally the version values get written via a plugin specifically for
Android. While this could be done manually (with proper Ruby knowledge),
this keeps it simple and focused on the main task.